refactor(types): move MCP catalog & OAuth template schema to librefang-types
Honors the librefang-types schema-spine convention from
crates/librefang-types/CLAUDE.md ("Owns: every cross-crate type"). The
moved types had been living in `librefang-extensions/src/lib.rs` even
though they were imported by `librefang-api`, `librefang-cli`, and
re-exposed through kernel APIs.

Migrated:
- McpCategory, McpCatalogTransport, McpCatalogRequiredEnv,
  McpCatalogEntry, McpCatalogI18n, McpStatus
  -> librefang_types::mcp
- OAuthTemplate -> librefang_types::oauth (sibling of existing
  OAuthTokens, which already lives there)

Renamed HealthCheckConfig -> McpHealthCheckConfig to disambiguate from
the pre-existing librefang_types::config::HealthCheckConfig (global
LLM-provider health-check tuning, distinct two-field struct).

No re-export shim left in librefang-extensions per CLAUDE.md "avoid
backwards-compatibility hacks" rule -- all callers updated to import
from librefang_types::{mcp, oauth} directly. Behaviour-bearing modules
(catalog loading, installer transforms, health monitoring, OAuth PKCE
flow) stay in librefang-extensions; this change is types-only.

Refs: #3295
