+++ crates/librefang-types/src/lib.rs
@@ -17,6 +17,7 @@ pub mod event;
 pub mod goal;
 pub mod i18n;
 pub mod manifest_signing;
+pub mod mcp;
 pub mod media;
 pub mod memory;
 pub mod message;
